% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_baselers.R
\name{write_baselers}
\alias{write_baselers}
\title{write baselers data to file}
\usage{
write_baselers(nsim = 10000, seed = 1)
}
\arguments{
\item{nsim}{number of simulations}

\item{seed}{seed}
}
\description{
write baselers data to file
}
